🐛 FIX: Restart test app if coldbox bootstrap failed
This works around an issue where an error on application start (such as a datasource connection issue) prevents the `onApplicationStart()` method from firing. When this happens, the test app is left in an unloaded state and ColdBox is not ready for testing.

To fix, we simply re-run `onApplicationStart()` if the bootstrap class is not found.